refactor(ports): give the DocumentIndex double and its suite one home
Preparation for the tool rewiring, which needs a working index in
server-core's tests. mcp-server owned the only one, server-core cannot
import a composition root, and writing a second would be a third
implementation of one rule held to nothing — the shape this whole line of
work exists to remove.

So the in-memory implementation and the conformance suite move to
canvas-ports/test-utils behind a `./test-utils` export. The suite was
written as a factory for exactly this, and the deferral was explicit: it
would move when a second implementation appeared. Two have.

generateCanvasId moves with them, from server-core down to canvas-model.
canvas-ports cannot import a package above it, and the model is where the
ULID's SHAPE already lives — having the shape and the minting in one place
is the same single-sourcing as the path rule.

One contract gap the wiring surfaced: listDocuments answered an unknown
workspace with an empty list, but wbCanvasList's rule is that a
never-persisted workspace is an error — a typo'd workspaceId is otherwise
indistinguishable from a workspace that genuinely holds nothing. It now
fails WorkspaceNotFoundError, mutation-checked in both implementations.
