+++ packages/canvas-model/src/index.ts
@@ -1,13 +1,13 @@
 export * from './asset-ref.js'
 export * from './clipboard.js'
 export * from './facets.js'
+// The mdast subset is intentionally NOT re-exported here — it is
+// versioned, reached via the package's `./mdast` subpath export instead of
+// the stable public surface.
+export { generateCanvasId } from './generate-canvas-id.js'
 export * from './ids.js'
 export * from './json-schema.js'
 export * from './markdown.js'
 export * from './meta.js'
 export * from './spatial.js'
 export * from './workspace-tree.js'
-
-// The mdast subset is intentionally NOT re-exported here — it is
-// versioned, reached via the package's `./mdast` subpath export instead of
-// the stable public surface.
