+++ packages/canvas-ports/package.json
@@ -12,7 +12,11 @@
       "types": "./src/index.ts",
       "import": "./src/index.ts"
     },
-    "./package.json": "./package.json"
+    "./package.json": "./package.json",
+    "./test-utils": {
+      "types": "./src/test-utils/index.ts",
+      "import": "./src/test-utils/index.ts"
+    }
   },
   "scripts": {
     "typecheck": "tsc --noEmit",
