+++ packages/canvas-ports/src/test-utils/in-memory-document-index.test.ts
@@ -1,9 +1,9 @@
 import { describe } from 'vitest'
-import { describeDocumentIndexConformance } from '../document-index-conformance.js'
+import { describeDocumentIndexConformance } from './document-index-conformance.js'
 import { InMemoryDocumentIndex } from './in-memory-document-index.js'
 
-// The same suite the sqlite store answers to. Two implementations is what
-// makes it a conformance suite rather than one store's tests.
+// The same suite the sqlite store answers to, run here so the double cannot
+// drift from the guarantees while still satisfying the interface.
 describe('InMemoryDocumentIndex', () => {
   describeDocumentIndexConformance(async () => ({
     index: new InMemoryDocumentIndex(),
