+++ packages/canvas-ports/src/test-utils/index.ts
@@ -1,0 +1,2 @@
+export { describeDocumentIndexConformance } from './document-index-conformance.js'
+export { InMemoryDocumentIndex } from './in-memory-document-index.js'
