+++ packages/mcp-server/src/server/store/inmemory/index.ts
@@ -1,3 +1,2 @@
 export { InMemoryBlobStore } from './in-memory-blob-store.js'
 export { InMemoryCanvasDocStore } from './in-memory-canvas-doc-store.js'
-export { InMemoryDocumentIndex } from './in-memory-document-index.js'
