+++ packages/server-core/src/index.ts
@@ -86,7 +86,6 @@ export {
 } from './tools/export-okf.js'
 export type { FacetSetInput, FacetSetOutput } from './tools/facet-set.js'
 export { createFacetSetTool, facetSetInputSchema, facetSetOutputSchema } from './tools/facet-set.js'
-export { generateCanvasId } from './tools/generate-canvas-id.js'
 export type { NodeAddInput, NodeAddOutput } from './tools/node-add.js'
 export { createNodeAddTool, nodeAddInputSchema, nodeAddOutputSchema } from './tools/node-add.js'
 export type { NodeLockInput, NodeLockOutput } from './tools/node-lock.js'
